+++ connection.cpp
@@ -68,6 +68,43 @@ QPainterPath Connection::shape() const
     return line_;
 }
 
+void Connection::updateShape()
+{
+    prepareGeometryChange();
+
+    switch (data_.cordType()) {
+    case ConnectionCordType::Linear: {
+        line_.clear();
+        line_.moveTo(pt0_);
+        line_.lineTo(pt1_);
+
+        QPainterPathStroker stroker;
+        stroker.setWidth(pen_width_);
+        stroker.setCapStyle(Qt::RoundCap);
+        line_ = stroker.createStroke(line_);
+    } break;
+    case ConnectionCordType::Bezier: {
+        auto bezy = (std::abs(pt0_.x() - pt1_.x()) < 20) ? 20 : 40;
+        ctl_pt0_ = pt0_ + QPointF(0, bezy);
+        ctl_pt1_ = pt1_ + QPointF(0, -bezy);
+        line_.clear();
+        line_.moveTo(pt0_);
+        line_.cubicTo(ctl_pt0_, ctl_pt1_, pt1_);
+        line_.cubicTo(ctl_pt1_, ctl_pt0_, pt0_);
+        line_.closeSubpath();
+
+        QPainterPathStroker stroker;
+        stroker.setWidth(pen_width_);
+        line_ = stroker.createStroke(line_);
+    } break;
+    case ConnectionCordType::Segmented:
+    default:
+        break;
+    }
+
+    update(boundingRect());
+}
+
 QRectF Connection::boundingRect() const
 {
     return line_.controlPointRect();
@@ -85,34 +122,24 @@ XletInfo Connection::sourceInfo() const
 
 void Connection::setPoints(const QPointF& p0, const QPointF& p1)
 {
-    prepareGeometryChange();
-
-    line_.clear();
-    line_.moveTo(p0);
-
-    auto bezy = (std::abs(p0.x() - p1.x()) < 20) ? 20 : 40;
-
     pt0_ = p0;
     pt1_ = p1;
-    ctl_pt0_ = pt0_ + QPointF(0, bezy);
-    ctl_pt1_ = pt1_ + QPointF(0, -bezy);
 
-    line_.cubicTo(ctl_pt0_, ctl_pt1_, pt1_);
-    line_.cubicTo(ctl_pt1_, ctl_pt0_, pt0_);
-    line_.closeSubpath();
-
-    QPainterPathStroker stroker;
-    stroker.setWidth(pen_width_);
-    line_ = stroker.createStroke(line_);
-
-    update(boundingRect());
+    updateShape();
 }
 
 void Connection::setStyle(ConnectionStyle style)
 {
     pen_width_ = ConnectionStyleDatabase::instance().penWidth(style, 1.5);
     color_ = ConnectionStyleDatabase::instance().color(style, Qt::darkGray);
-    update(boundingRect());
+
+    updateShape();
+}
+
+void Connection::setCordType(ConnectionCordType type)
+{
+    data_.setCordType(type);
+    updateShape();
 }
 
 void Connection::contextMenuEvent(QGraphicsSceneContextMenuEvent* event)
@@ -122,11 +149,30 @@ void Connection::contextMenuEvent(QGraphicsSceneContextMenuEvent* event)
         return;
 
     QMenu menu;
-    auto act = menu.addAction(QAction::tr("Delete"));
-    QAction::connect(act, &QAction::triggered, dia_scene,
+
+    auto act_del = menu.addAction(QAction::tr("Delete"));
+    QAction::connect(act_del, &QAction::triggered, dia_scene,
         [this, dia_scene]() {
             emit dia_scene->removeConnection(data_);
         });
+
+    auto menu_ct = menu.addMenu(QAction::tr("Cord type"));
+
+    auto act_bezier = menu_ct->addAction(QAction::tr("Bezier"));
+    auto act_linear = menu_ct->addAction(QAction::tr("Linear"));
+
+    act_bezier->setCheckable(true);
+    act_bezier->setChecked(data_.cordType() == ConnectionCordType::Bezier);
+    QAction::connect(act_bezier, &QAction::triggered, dia_scene, [this]() {
+        setCordType(ConnectionCordType::Bezier);
+    });
+
+    act_linear->setCheckable(true);
+    act_linear->setChecked(data_.cordType() == ConnectionCordType::Linear);
+    QAction::connect(act_linear, &QAction::triggered, dia_scene, [this]() {
+        setCordType(ConnectionCordType::Linear);
+    });
+
     menu.exec(event->screenPos());
     event->accept();
 }